Tests: rename test folder to tests to exclude for lcov result
By default tests folder is not included into lcov result. To make result
more accurate, remove tests.
