+++ example/ExampleBackend.hpp
@@ -27,7 +27,7 @@
 #include <xen/be/FrontendHandlerBase.hpp>
 #include <xen/be/RingBufferBase.hpp>
 
-#include "../test/testProtocol.h"
+#include "tests/testProtocol.h"
 
 //! [ExampleOutRingBuffer]
 class ExampleOutRingBuffer : public XenBackend::RingBufferOutBase
