+++ tests/mocks/XenCtrlMock.cpp
