+++ tests/mocks/XenStoreMock.cpp
