+++ tests/testRingBuffer.cpp
@@ -24,7 +24,7 @@
 #include <condition_variable>
 #include <mutex>
 
-#include <catch.hpp>
+#include "catch.hpp"
 
 #include "mocks/XenEvtchnMock.hpp"
 #include "mocks/XenGnttabMock.hpp"
