+++ tests/testXenEvtchn.cpp
@@ -22,7 +22,7 @@
 #include <condition_variable>
 #include <mutex>
 
-#include <catch.hpp>
+#include "catch.hpp"
 
 #include "mocks/XenEvtchnMock.hpp"
 #include "XenEvtchn.hpp"
